backend/requirements.txt
# Financial Data ML System Requirements
# Core Python packages for data collection and analysis

# Data manipulation and analysis
pandas>=1.5.0
numpy>=1.21.0

# HTTP requests and API interactions
requests>=2.28.0
urllib3>=1.26.0
flask>=2.0.0
flask-cors>=3.0.10
python-dotenv>=1.0.0

# Date and time handling (additional utilities)
python-dateutil>=2.8.0

# Data validation and processing
pydantic>=1.10.0

# Financial data sources
yfinance>=0.2.0

# Note: sqlite3 is built into Python 3.x, no need to install separately

# Logging and configuration
configparser>=5.0.0

# Math and statistics
scipy>=1.9.0
scikit-learn>=1.1.0

# Optional: For enhanced data processing
# openpyxl>=3.0.0  # Excel file support
matplotlib>=3.5.0  # Plotting (for future analysis)
# seaborn>=0.11.0  # Statistical plotting (for future analysis)

# Development and testing (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0
# mypy>=0.990

# Documentation (optional)
# sphinx>=5.0.0
# sphinx-rtd-theme>=1.0.0

# Production WSGI server
waitress>=2.1.2